코드 리팩터링
