+++ app/src/main/java/carehalcare/carehalcare/Feature_write/DividerItemDecorator.java
@@ -1,0 +1,28 @@
+package carehalcare.carehalcare.Feature_write;
+
+import android.graphics.Canvas;
+import android.graphics.drawable.Drawable;
+import android.view.View;
+
+import androidx.recyclerview.widget.RecyclerView;
+
+public class DividerItemDecorator extends RecyclerView.ItemDecoration {
+    private Drawable mDivider;
+    public DividerItemDecorator(Drawable divider) { mDivider = divider; }
+
+    @Override
+    public void onDraw(Canvas canvas, RecyclerView parent, RecyclerView.State state) {
+        int dividerLeft = parent.getPaddingLeft();
+        int dividerRight = parent.getWidth() - parent.getPaddingRight();
+        int childCount = parent.getChildCount();
+        for(int i=0;i<=childCount-2; i++){
+            View child = parent.getChildAt(i);
+            RecyclerView.LayoutParams params = (RecyclerView.LayoutParams)child.getLayoutParams();
+            int dividerTop = child.getBottom()+params.bottomMargin;
+            int dividerBottom = dividerTop+mDivider.getIntrinsicHeight();
+            mDivider.setBounds(dividerLeft, dividerTop, dividerRight, dividerBottom);
+            mDivider.draw(canvas);
+        }
+    }
+}
+
